nearly done

+++ linearRegression.py
@@ -11,14 +11,15 @@ class LinearRegression:
         
     def fit(self, X, y):
         n_samples, n_features = X.shape
-        self.weights = np.zeros(n_features)
+        self.weights = np.zeros(n_features).astype(np.float64)
         self.bias = 0
         
         for i in range(self.n_iterations):
             y_predicted = np.dot(X, self.weights) + self.bias
             dw = (1/n_samples) * np.dot(X.T, (y_predicted - y))
             db = (1/n_samples) * np.sum(y_predicted - y)
-            self.weights -= self.learning_rate * dw
+           
+            self.weights -= (self.learning_rate * dw).astype(np.float64)
             self.bias -= self.learning_rate * db
             cost = (1/n_samples) * np.sum((y_predicted - y)**2)
             self.cost_history.append(cost)
